requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
colorama==0.4.6
decorator==5.1.1
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-extensions==3.2.3
django-intl-tel-input==0.3.1
django-phonenumber-field==7.1.0
exceptiongroup==1.1.3
executing==1.2.0
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
phonenumbers==8.13.17
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
python-decouple==3.8
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
